Fixed the --warnings_as_errors output.
As @ottojo points in #1 this option counted warnings as errors, but
never printed the warnings.
Now the code does what the option says: warnings are converted to
errors. Not just counted as errors, but also printed as errors.
I think this is more coherent with the name of the option.